Extract sending UDP & TCP queries to module
